Doc: Also fix qdoc signature of qHash(const QTypeRevision &key ...)
Amends c6cdf38e752c22babdbe645366bdfb7ce51d01ff

+++ src/corelib/tools/qversionnumber.cpp
@@ -739,7 +739,7 @@ QDebug operator<<(QDebug debug, const QTypeRevision &revision)
 #endif
 
 /*!
-    \fn uint qHash(const QTypeRevision &key, uint seed)
+    \fn size_t qHash(const QTypeRevision &key, size_t seed)
     \relates QHash
     \since 6.0
